# frontend/api_documentation.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Price Tracker API",
    "description": "\n    ## API REST para rastrear precios de productos en línea\n    \n    Sistema multi-usuario con autenticación JWT que permite:\n    - Rastrear precios de productos de diferentes tiendas en línea\n    - Historial completo de cambios de precios\n    - Alertas cuando se alcanza el precio objetivo\n    - Gráficos de evolución de precios\n    - Soporte para múltiples sitios web (MercadoLibre, Amazon, eBay, etc.)\n    \n    ### Autenticación\n    Esta API utiliza JWT (JSON Web Tokens) para autenticación. Para usar los endpoints protegidos:\n    1. Registra un usuario en `/api/auth/register`\n    2. Inicia sesión en `/api/auth/login` para obtener tu token\n    3. Usa el token en el header: `Authorization: Bearer {token}`\n    \n    ### Tecnologías\n    - FastAPI 0.104.1\n    - PostgreSQL (Supabase)\n    - SQLAlchemy 2.0.44\n    - JWT Authentication\n    - bcrypt password hashing\n    - BeautifulSoup4 web scraping\n    \n    ### Desarrollador\n    **HellSpawn**\n    \n    Sistema desarrollado como proyecto profesional de rastreo de precios con arquitectura moderna y escalable.\n    \n    ### Enlaces\n    - Documentación interactiva: `/docs`\n    - Documentación alternativa: `/redoc`\n    - GitHub: [Próximamente]\n    ",
    "contact": {
      "name": "HellSpawn",
      "email": "hellspawn@pricetracker.dev"
    },
    "license": {
      "name": "MIT License",
...
    },
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Register",
        "description": "Registrar nuevo usuario\n\n- **email**: Email único del usuario\n- **username**: Nombre de usuario único (3-50 caracteres)\n- **password**: Contraseña (mínimo 6 caracteres)",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Login",
        "description": "Iniciar sesión y obtener token JWT\n\n- **username**: Nombre de usuario o email\n- **password**: Contraseña",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_api_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login/json": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Login Json",
        "description": "Iniciar sesión con JSON (alternativa a form-data)\n\n- **username**: Nombre de usuario o email\n- **password**: Contraseña",
        "operationId": "login_json_api_auth_login_json_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Get Current User Info",
        "description": "Obtener información del usuario actual\nRequiere token JWT válido",
        "operationId": "get_current_user_info_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/auth/me/deactivate": {
      "put": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Deactivate Account",
        "description": "Desactivar cuenta de usuario",
        "operationId": "deactivate_account_api_auth_me_deactivate_put",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_api_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_api_auth_login_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "user"
        ],
        "title": "Token"
      },
      "User": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "email",
          "username",
          "id",
          "is_active",
          "created_at"
        ],
        "title": "User"
      },
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 100,
            "minLength": 6,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "username",
          "password"
        ],
        "title": "UserCreate"
      },
      "UserLogin": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserLogin"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}